+++ src/gui-qt4/libs/seiscomp3/gui/datamodel/eventsummary.cpp
@@ -11,8 +11,6 @@
  ***************************************************************************/
 
 
-
-
 #include <seiscomp3/gui/datamodel/eventsummary.h>
 
 #include <seiscomp3/core/system.h>