+++ src/gui-qt4/libs/seiscomp3/gui/datamodel/eventsummaryview.cpp
@@ -11,8 +11,6 @@
  ***************************************************************************/
 
 
-
-
 #define SEISCOMP_COMPONENT EventSummaryView
 
 #include "eventsummaryview.h"